requirements.txt
python-telegram-bot[job-queue]==20.7
apscheduler==3.10.4
python-dotenv==1.0.0
Pillow==10.0.1
